Fix opcode struct Spell::SendSpellStart, LootView::WriteData,    HandleUpdateAccountData. Fix LoadDBC for loading data from DB.

+++ src/server/game/Spells/Spell.cpp
@@ -4186,8 +4186,8 @@ void Spell::SendSpellStart()
             data << uint8(powerType);
         }*/
 
-        data << uint8(m_spellInfo->PowerType);
-        data << int32(m_caster->GetPower((Powers)m_spellInfo->PowerType));
+        data << uint32(m_spellInfo->PowerType);
+        data << uint32(m_caster->GetPower((Powers)m_spellInfo->PowerType));
     }
 
     data << uint32(castFlags);
